pyproject.toml
[project]
name = "kinect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "av>=16.0.1",
    "camera>=1.3.0",
    "cv2-enumerate-cameras>=1.3.0",
    "freenect>=0.1.0",
    "opencv-python>=4.12.0.88",
    "pykinect2>=0.1.0",
    "scikit-image>=0.25.2",
]
